# Optimizations for toArray and fromArray